switch from double to float to keep 32 bit width in std::variant for verifydtapn engine

+++ include/PQL/Expressions.h
@@ -119,7 +119,7 @@ namespace unfoldtacpn {
         };
 
         using LiteralIntExpr = LiteralExpr<int>;
-        using LiteralRealExpr = LiteralExpr<double>;
+        using LiteralRealExpr = LiteralExpr<float>;
 
         class IdentifierExpr : public Expr {
         public:
